requirements.txt
 streamlit
 langchain 
 openai 
 pypdf 
 faiss-cpu 
 langchain_community